+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Moved some `dtypes` support functionality from GoMLX to Gopjrt. 
 * Added BFloat16 alias.
 * Renamed `FromGoType` to `FromGenericsType` and `FromType` to `FromGoType`, to maintain naming consistency.
+* Added DType.Memory as an alias to DType.Size.
 
 # v0.1.2 SuppressAbseilLoggingHack